requirements.txt
3==0.0.0
aiodns==3.1.1
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
ccxt==4.2.44
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
install==1.3.5
magic-filter==1.0.12
multidict==6.0.5
peewee==3.17.1
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
pybit==5.6.2
pycares==4.4.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.5.3
pydantic-settings==2.2.1
pydantic_core==2.14.6
python-dotenv==1.0.1
requests==2.31.0
setuptools==69.1.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
urllib3==2.2.0
websocket-client==1.7.0
websockets==12.0
yarl==1.9.4
